+++ packages/admin/src/Http/Livewire/Components/Settings/Currencies/CurrencyCreate.php
@@ -17,6 +17,13 @@ class CurrencyCreate extends Component
      */
     public Currency $currency;
 
+    /**
+     * Determine whether to show format info text.
+     *
+     * @var bool
+     */
+    public $showFormatInfo = false;
+
     public function mount()
     {
         $this->currency = new Currency();
@@ -33,9 +40,6 @@ class CurrencyCreate extends Component
             'currency.code'           => 'required|max:255|unique:'.$this->currency->getTable().',code',
             'currency.name'           => 'required|max:255',
             'currency.exchange_rate'  => 'required|numeric|min:0.1|max:255',
-            'currency.format'         => 'required|max:255|regex:/{value}/i',
-            'currency.decimal_point'  => 'required|max:255',
-            'currency.thousand_point' => 'required|max:255',
             'currency.decimal_places' => 'required|integer|max:4',
             'currency.enabled'        => 'nullable',
             'currency.default'        => 'nullable',