+++ packages/admin/tests/TestCase.php
@@ -5,10 +5,10 @@ namespace GetCandy\Hub\Tests;
 use Cartalyst\Converter\Laravel\ConverterServiceProvider;
 use GetCandy\GetCandyServiceProvider;
 use GetCandy\Hub\AdminHubServiceProvider;
+use GetCandy\Tests\Stubs\TestUrlGenerator;
 use Illuminate\Support\Facades\Config;
 use Kalnoy\Nestedset\NestedSetServiceProvider;
 use Livewire\LivewireServiceProvider;
-use LivewireUI\Modal\LivewireModalServiceProvider;
 use Spatie\Activitylog\ActivitylogServiceProvider;
 use Spatie\MediaLibrary\MediaLibraryServiceProvider;
 
@@ -21,6 +21,7 @@ class TestCase extends \Orchestra\Testbench\TestCase
         Config::set('auth.guards.staff', [
             'driver' => 'getcandyhub',
         ]);
+        Config::set('getcandy.urls.generator', TestUrlGenerator::class);
     }
 
     protected function getPackageProviders($app)
@@ -30,7 +31,6 @@ class TestCase extends \Orchestra\Testbench\TestCase
             LivewireServiceProvider::class,
             AdminHubServiceProvider::class,
             ActivitylogServiceProvider::class,
-            LivewireModalServiceProvider::class,
             MediaLibraryServiceProvider::class,
             ConverterServiceProvider::class,
             NestedSetServiceProvider::class,