+++ packages/admin/tests/Unit/Http/Livewire/Components/Products/ProductShowTest.php
@@ -284,7 +284,9 @@ class ProductShowTest extends TestCase
             ->test(ProductShow::class, [
                 'product' => $product->refresh(),
             ])->set('attributeMapping.'.'a_'.$name->id.'.data', 'nouseforaname')
-            ->set('attributeMapping.'.'a_'.$description->id.'.data', 'nouseforadescription');
+            ->set('attributeMapping.'.'a_'.$description->id.'.data', 'nouseforadescription')
+            ->call('addUrl')
+            ->set('urls.0.slug', 'foo-bar');
 
         $component->call('save')->assertHasNoErrors();
 
@@ -435,6 +437,8 @@ class ProductShowTest extends TestCase
             ->test(ProductShow::class, [
                 'product' => $product,
             ])->set('optionValues', $values->pluck('id')->toArray())
+            ->call('addUrl')
+            ->set('urls.0.slug', 'foo-bar')
             ->call('save')
             ->assertHasNoErrors();
 
@@ -549,7 +553,9 @@ class ProductShowTest extends TestCase
         $component = LiveWire::actingAs($staff, 'staff')
             ->test(ProductShow::class, [
                 'product' => $product->refresh(),
-            ])->assertCount('collections', 0)
+            ])->call('addUrl')
+            ->set('urls.0.slug', 'foo-bar')
+            ->assertCount('collections', 0)
             ->set('collections', collect([[
                 'id' => $collection->id,
                 'name' => $collection->translateAttribute('name'),
@@ -608,7 +614,9 @@ class ProductShowTest extends TestCase
         LiveWire::actingAs($staff, 'staff')
             ->test(ProductShow::class, [
                 'product' => $product->refresh(),
-            ])->assertCount('associations', 0)
+            ])->call('addUrl')
+            ->set('urls.0.slug', 'foo-bar')
+            ->assertCount('associations', 0)
             ->set('associations', collect([
                 [
                     'inverse' => false,
@@ -669,7 +677,9 @@ class ProductShowTest extends TestCase
         LiveWire::actingAs($staff, 'staff')
             ->test(ProductShow::class, [
                 'product' => $product->refresh(),
-            ])->assertCount('associations', 0)
+            ])->call('addUrl')
+            ->set('urls.0.slug', 'foo-bar')
+            ->assertCount('associations', 0)
             ->set('associations', collect([
                 [
                     'inverse' => true,
@@ -685,5 +695,124 @@ class ProductShowTest extends TestCase
             'product_target_id' => $product->id,
             'type' => 'cross-sell',
         ]);
+    }
+
+    /**
+     * @test
+     * */
+    public function variants_are_enabled_by_default()
+    {
+        $staff = Staff::factory()->create([
+            'admin' => true,
+        ]);
+
+        $product = Product::factory()->create([
+            'status' => 'published',
+            'brand'  => 'BAR',
+        ]);
+
+        $variant = ProductVariant::factory()->create([
+            'product_id' => $product->id,
+        ]);
+
+        foreach (Currency::get() as $currency) {
+            Price::factory()->create([
+                'priceable_type' => ProductVariant::class,
+                'priceable_id'   => $variant->id,
+                'currency_id'    => $currency->id,
+                'tier'           => 1,
+            ]);
+        }
+
+        LiveWire::actingAs($staff, 'staff')
+            ->test(ProductShow::class, [
+                'product' => $product,
+            ])->assertSeeHtml('Variants');
+    }
+
+    /**
+     * @test
+     * */
+    public function variants_can_be_disabled()
+    {
+        Config::set('getcandy-hub.products.disable_variants', true);
+
+        $staff = Staff::factory()->create([
+            'admin' => true,
+        ]);
+
+        $product = Product::factory()->create([
+            'status' => 'published',
+            'brand'  => 'BAR',
+        ]);
+
+        $variant = ProductVariant::factory()->create([
+            'product_id' => $product->id,
+        ]);
+
+        foreach (Currency::get() as $currency) {
+            Price::factory()->create([
+                'priceable_type' => ProductVariant::class,
+                'priceable_id'   => $variant->id,
+                'currency_id'    => $currency->id,
+                'tier'           => 1,
+            ]);
+        }
+
+        LiveWire::actingAs($staff, 'staff')
+            ->test(ProductShow::class, [
+                'product' => $product,
+            ])->assertDontSeeHtml('Variants');
+    }
+
+    /**
+     * @test
+     * */
+    public function variants_arent_generated_when_disabled()
+    {
+        Config::set('getcandy-hub.products.disable_variants', true);
+
+        $staff = Staff::factory()->create([
+            'admin' => true,
+        ]);
+
+        $product = Product::factory()->create([
+            'status' => 'published',
+            'brand'  => 'BAR',
+        ]);
+
+        $variant = ProductVariant::factory()->create([
+            'product_id' => $product->id,
+        ]);
+
+        foreach (Currency::get() as $currency) {
+            Price::factory()->create([
+                'priceable_type' => ProductVariant::class,
+                'priceable_id'   => $variant->id,
+                'currency_id'    => $currency->id,
+                'tier'           => 1,
+            ]);
+        }
+
+        ProductOption::factory(2)->create()->each(function ($option) {
+            $option->values()->createMany(
+                ProductOptionValue::factory(2)->make()->toArray()
+            );
+        });
+
+        $values = ProductOptionValue::get();
+
+        Config::set('getcandy-hub.products.sku.unique', true);
+
+        LiveWire::actingAs($staff, 'staff')
+            ->test(ProductShow::class, [
+                'product' => $product,
+            ])->call('addUrl')
+            ->set('urls.0.slug', 'foo-bar')
+            ->set('optionValues', $values->pluck('id')->toArray())
+            ->call('save')
+            ->assertHasNoErrors();
+
+        $this->assertEquals(1, $product->variants()->count());
     }
 }