+++ packages/core/src/Console/Commands/MeilisearchSetup.php
@@ -2,10 +2,6 @@
 
 namespace GetCandy\Console\Commands;
 
-use GetCandy\Models\Customer;
-use GetCandy\Models\Order;
-use GetCandy\Models\Product;
-use GetCandy\Models\ProductOption;
 use Illuminate\Console\Command;
 use Laravel\Scout\EngineManager;
 use Laravel\Scout\Engines\MeiliSearchEngine;
@@ -30,58 +26,47 @@ class MeilisearchSetup extends Command
     protected MeiliSearchEngine $engine;
 
     /**
-     * The models we want to search on.
-     *
-     * @var array
-     */
-    protected $searchables = [
-        Product::class,
-        Order::class,
-        ProductOption::class,
-        Customer::class,
-    ];
-
-    /**
      * Execute the console command.
      *
      * @return mixed
      */
     public function handle(EngineManager $engine)
     {
+        // Return the models we want to search on.
+        $searchables = config('getcandy.search.models', []);
+
         $this->engine = $engine->createMeilisearchDriver();
 
         // Make sure we have the relevant indexes ready to go.
-        foreach ($this->searchables as $searchable) {
-            $indexName = (new $searchable())->searchableAs();
+        foreach ($searchables as $searchable) {
+            $model = (new $searchable());
+
+            $indexName = $model->searchableAs();
 
             try {
                 $index = $this->engine->getIndex($indexName);
+                $this->warn("Index {$indexName} found for {$searchable}");
             } catch (ApiException $e) {
-                $this->info("Creating index for {$searchable}");
+                $this->warn($e->getMessage());
+                $this->info("Creating index {$indexName} for {$searchable}");
                 $this->engine->createIndex($indexName);
-
+                sleep(1);
                 $index = $this->engine->getIndex($indexName);
             }
 
-            $index->updateFilterableAttributes([
-                '__soft_deleted',
-            ]);
+            $this->info("Adding filterable fields to {$searchable}");
+
+            $index->updateFilterableAttributes(
+                $model->getFilterableAttributes()
+            );
+
+            $this->info("Adding sortable fields to {$searchable}");
+
+            $index->updateSortableAttributes(
+                $model->getSortableAttributes()
+            );
+
+            $this->newLine();
         }
-
-        $this->setUpOrders();
-    }
-
-    protected function setUpOrders()
-    {
-        $index = $this->engine->getIndex(
-            (new Order())->searchableAs()
-        );
-
-        $index->updateFilterableAttributes([
-            'status',
-            'placed_at',
-            'created_at',
-            'total',
-        ]);
     }
 }